Remove model.py
